add databaseConfing
